some changes

+++ src/app/announcer-page/announcer-page.component.ts
@@ -54,21 +54,24 @@ export class AnnouncerPageComponent implements OnInit {
     this.showFileModal = false
   }
 
-  deleteAnnouncer(id: number) {
-    this.request.deleteAnnouncerLots(id).subscribe(response => {
-      location.reload()
-    }, error => {
-      if(error.status == '401') {
-        this.request.refreshToken().subscribe( (response: any) =>  {
-          localStorage.setItem('access_token', response.access_token)
-          location.reload()
-        }, errorToken => {
-          alert(errorToken.message)
-        })
-      } else {
-        alert(error.message)
-      }
-    })
+  deleteAnnouncer(id: number, name: string) {
+    let quest = confirm(`Вы хотите удалить ${name}`)
+    if(quest == true) {
+      this.request.deleteAnnouncerLots(id).subscribe(response => {
+        location.reload()
+      }, error => {
+        if(error.status == '401') {
+          this.request.refreshToken().subscribe( (response: any) =>  {
+            localStorage.setItem('access_token', response.access_token)
+            location.reload()
+          }, errorToken => {
+            alert(errorToken.message)
+          })
+        } else {
+          alert(error.message)
+        }
+      })
+    }
   }
 
 }
